feat: pss-cln-rec-contract-roots-004 - Fold replay root cluster into private replay
