+++ cmd/packagetargetmanifestcheck/recordings_root_contract_test.go
@@ -26,7 +26,6 @@ func TestRecordingsExcessRootContractFoldDestinations(t *testing.T) {
 
 	want := map[string]string{
 		"artifacts":             "pkg/services/recordings/internal/services/artifacts_export",
-		"replay":                "pkg/services/recordings/internal/services/replay",
 		"live_recording_target": "pkg/services/recordings/internal/services/recording_lifecycle",
 	}
