+++ pkg/services/factory_definitions/owned_contract.go
@@ -14,7 +14,8 @@
 // (canonical_ledger/event_contract.go with peer aliases in contracts.go). World-state,
 // dispatch, and workstation-request vocabulary is owned by projection_query with peer
 // aliases in recordings/contracts.go. Replay vocabulary is owned by
-// pkg/services/recordings/replay (replay_contract.go). Worker execution vocabulary is
+// recordings/internal/services/replay (replay_contract.go with peer aliases in
+// contracts.go). Worker execution vocabulary is
 // owned by pkg/services/workers (worker_vocabulary_contract.go,
 // execution_contracts.go); provider-session identity is published at
 // pkg/services/providers (SessionRef) and pkg/services/workers
